♻️ [Refactoring]: [BE] parent_validation.rs を廃止し TaskIndex aggregate に validation を統合
DDD 原則「validation は専用ファイルではなくドメインオブジェクトに紐づける」
に従い、parent_validation.rs を廃止して中身を task_index.rs に移設した。
create/validate.rs を廃止した際と同じ方針。

- ParentHierarchyErrorReason / ParentValidationFailure / MAX_PARENT_DEPTH を
  task_index.rs に同居
- validate_parent_existence / validate_parent_hierarchy / resolve_parent_for_new_task
  / validate_chain_from_parent を task_index.rs 内の pub(super) helper として配置
  （TaskIndex aggregate の不変条件を守る sibling helper）
- 親チェーン検証用テストは task_index_parent_chain_tests.rs に rename
  (#[path] 経由で task_index.rs から参照)
- children.rs / parse.rs / create/error.rs の import を task_index 起点に変更
- task.rs から pub mod parent_validation を削除
- CLAUDE.md / docs/impl/ddd-domain-design.md を新構造に追従

公開振る舞いと既存テスト関数名は不変。
