+++ src-tauri/src/task/children.rs
@@ -2,10 +2,9 @@
 //!
 //! parent 検証を委譲呼び出しした上で、parent 逆引きで各 Task の `children` を埋める。
 
-use crate::task::parent_validation::validate_parent_hierarchy;
 use crate::task::parse::TaskParseError;
 use crate::task::path_lookup::{append_child_to_parent, clear_children, task_lookup_index};
-use crate::task::task_index::Task;
+use crate::task::task_index::{validate_parent_hierarchy, Task};
 
 /// 全 Task の parent 参照を検証し、親 Task の children を parent 逆引きで構築する。
 pub(super) fn build_children(tasks: Vec<Task>) -> Result<Vec<Task>, TaskParseError> {
