+++ src-tauri/src/task/parse.rs
@@ -12,10 +12,9 @@ use crate::config::column_name::ColumnName;
 use crate::config::Config;
 use crate::task::frontmatter::{parse_bytes, FrontmatterError, Parsed};
 use crate::task::label::Label;
-use crate::task::parent_validation::ParentHierarchyErrorReason;
 use crate::task::path_normalization::normalize_path_parts;
 use crate::task::task_file_path::TaskFilePath;
-use crate::task::task_index::Task;
+use crate::task::task_index::{ParentHierarchyErrorReason, Task};
 use crate::task::task_title::TaskTitle;
 use crate::task::warning::{TaskWarning, TaskWarningCode};
