+++ src-tauri/src/task/task_index.rs
@@ -1,10 +1,12 @@
 //! Task aggregate ドメイン。
 //!
-//! `Task` entity と `TaskIndex` aggregate root を同居させる。`TaskIndex` は parent
-//! 検証 / children 派生 / reverse_links 派生といったドメインルールを sibling
-//! モジュールへ委譲し、外部からは aggregate メソッド経由でのみ振る舞いを呼び出す。
+//! `Task` entity と `TaskIndex` aggregate root を同居させる。`TaskIndex` の
+//! 不変条件（parent 存在 / 親チェーンに循環なし / 親チェーン深さ ≤ MAX）と
+//! それを検証するロジックは、DDD 戦術的パターンに従い aggregate root の責務
+//! としてこのファイルに集約する（独立した「validation」ファイルは作らない）。
 
 use std::collections::{BTreeMap, HashMap, HashSet};
+use std::fmt;
 use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};
@@ -13,19 +15,17 @@ use crate::config::column_name::ColumnName;
 use crate::task::children::build_children;
 use crate::task::frontmatter::Priority;
 use crate::task::label::Label;
-use crate::task::parent_validation::{
-    resolve_parent_for_new_task, validate_chain_from_parent, validate_parent_existence,
-    validate_parent_hierarchy, ParentHierarchyErrorReason, ParentValidationFailure,
-};
 use crate::task::parse::TaskParseError;
 use crate::task::path_lookup::{
     normalize_link_path_for_lookup, normalize_parent_path_for_lookup,
-    normalize_task_path_for_lookup,
+    normalize_task_path_for_lookup, parent_lookup_index, task_path_index,
 };
 use crate::task::reverse_links::build_reverse_links;
 use crate::task::task_file_path::TaskFilePath;
 use crate::task::task_title::TaskTitle;
-use crate::task::warning::TaskWarning;
+use crate::task::warning::{TaskWarning, TaskWarningCode};
+
+const MAX_PARENT_DEPTH: usize = 20;
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -51,8 +51,37 @@ pub struct Task {
     pub warnings: Vec<TaskWarning>,
 }
 
+/// 親チェーン違反の理由（循環 / 深さ超過）。
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ParentHierarchyErrorReason {
+    Cycle,
+    TooDeep,
+}
+
+impl fmt::Display for ParentHierarchyErrorReason {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Cycle => write!(f, "contains a cycle"),
+            Self::TooDeep => write!(f, "exceeds the maximum depth"),
+        }
+    }
+}
+
+/// 新規 task に対する `TaskIndex` の検証メソッドが返す失敗値。
+/// `task::create::error::CreateTaskError` から `From` で変換される。
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ParentValidationFailure {
+    NotFound {
+        parent: String,
+    },
+    ChainInvalid {
+        parent: String,
+        reason: ParentHierarchyErrorReason,
+    },
+}
+
 /// Task 集合の整合性（parent 存在 / 循環検出 / children・reverse_links 派生）を
-/// 守る Aggregate。
+/// 守る Aggregate。不変条件の検証メソッドを集約する。
 #[derive(Debug, Clone, PartialEq)]
 pub struct TaskIndex {
     tasks: Vec<Task>,
@@ -71,34 +100,40 @@ impl TaskIndex {
         &self.tasks
     }
 
+    /// parent 参照の存在のみを検証し、見つからない場合は warning を追加する。
     pub fn validate_parent_existence(self) -> Self {
         Self {
             tasks: validate_parent_existence(self.tasks),
         }
     }
 
+    /// parent 存在 + 循環 + 深さ検証を行う。
     pub fn validate_parent_hierarchy(self) -> Result<Self, TaskParseError> {
         Ok(Self {
             tasks: validate_parent_hierarchy(self.tasks)?,
         })
     }
 
+    /// 親検証を行ったうえで各 Task の `children` を逆引きで構築する。
     pub fn build_children(self) -> Result<Self, TaskParseError> {
         Ok(Self {
             tasks: build_children(self.tasks)?,
         })
     }
 
+    /// 各 Task の `links` を逆引きして `reverse_links` を構築する。
     pub fn build_reverse_links(self) -> Self {
         Self {
             tasks: build_reverse_links(self.tasks),
         }
     }
 
+    /// 新規 task が指す parent 文字列を既存 task 集合に対して解決する。
     pub fn resolve_parent_for_new_task(&self, parent: &str) -> Option<usize> {
         resolve_parent_for_new_task(parent, &self.tasks)
     }
 
+    /// 起点 parent から末端へ 1 edge 追加した chain の循環/深さ超過を検出する。
     pub fn validate_chain_from_parent(
         &self,
         parent_index: usize,
@@ -230,6 +265,137 @@ impl From<Vec<Task>> for TaskIndex {
     }
 }
 
+// ---------------------------------------------------------------------------
+// TaskIndex の不変条件を検証する sibling helper 群。
+//
+// `Vec<Task>` を直接受け取る形にしておくことで、aggregate を構築する前段の
+// 段階（open_project の scan 直後など）や、children/reverse_links 派生処理が
+// validate を委譲呼び出ししたい場面でも再利用できる。`pub(super)` で task
+// ドメイン内に閉じ、task ドメイン外からは TaskIndex aggregate のメソッド経由
+// でのみ呼び出す。
+// ---------------------------------------------------------------------------
+
+pub(super) fn validate_parent_existence(mut tasks: Vec<Task>) -> Vec<Task> {
+    let task_paths = task_path_index(&tasks);
+
+    for task in &mut tasks {
+        append_parent_not_found_warning(task, &task_paths);
+    }
+
+    tasks
+}
+
+pub(super) fn validate_parent_hierarchy(tasks: Vec<Task>) -> Result<Vec<Task>, TaskParseError> {
+    let tasks = validate_parent_existence(tasks);
+    let parent_lookup = parent_lookup_index(&tasks);
+
+    for task in &tasks {
+        validate_parent_chain(task, &parent_lookup)?;
+    }
+
+    Ok(tasks)
+}
+
+pub(super) fn resolve_parent_for_new_task(parent: &str, tasks: &[Task]) -> Option<usize> {
+    let normalized = normalize_parent_path_for_lookup(parent)?;
+    tasks
+        .iter()
+        .position(|task| normalize_task_path_for_lookup(task.file_path.as_str()) == normalized)
+}
+
+pub(super) fn validate_chain_from_parent(
+    parent_index: usize,
+    tasks: &[Task],
+) -> Result<(), ParentHierarchyErrorReason> {
+    let parent_task = tasks
+        .get(parent_index)
+        .expect("validate_chain_from_parent: parent_index must be in range (caller invariant)");
+    let lookup = parent_lookup_index(tasks);
+    let mut visited = HashSet::new();
+    let mut current = normalize_task_path_for_lookup(parent_task.file_path.as_str());
+    let mut depth: usize = 1;
+
+    loop {
+        if !visited.insert(current.clone()) {
+            return Err(ParentHierarchyErrorReason::Cycle);
+        }
+        if depth > MAX_PARENT_DEPTH {
+            return Err(ParentHierarchyErrorReason::TooDeep);
+        }
+        let Some(Some(next)) = lookup.get(&current) else {
+            return Ok(());
+        };
+        depth += 1;
+        current = next.clone();
+    }
+}
+
+fn validate_parent_chain(
+    task: &Task,
+    parent_lookup: &HashMap<String, Option<String>>,
+) -> Result<(), TaskParseError> {
+    let mut visited = HashSet::new();
+    let origin = task.file_path.as_str().to_string();
+    let mut current = normalize_task_path_for_lookup(task.file_path.as_str());
+    let mut depth = 0;
+
+    loop {
+        if !visited.insert(current.clone()) {
+            return Err(TaskParseError::CycleOrTooDeep {
+                file_path: origin,
+                reason: ParentHierarchyErrorReason::Cycle,
+            });
+        }
+
+        let Some(Some(parent)) = parent_lookup.get(&current) else {
+            return Ok(());
+        };
+
+        depth += 1;
+        if depth > MAX_PARENT_DEPTH {
+            return Err(TaskParseError::CycleOrTooDeep {
+                file_path: origin,
+                reason: ParentHierarchyErrorReason::TooDeep,
+            });
+        }
+
+        current = parent.clone();
+    }
+}
+
+fn append_parent_not_found_warning(task: &mut Task, task_paths: &HashSet<String>) {
+    let Some(parent) = &task.parent else {
+        return;
+    };
+
+    let Some(parent_lookup_path) = normalize_parent_path_for_lookup(parent.as_str()) else {
+        push_parent_not_found(task);
+        return;
+    };
+
+    if task_paths.contains(&parent_lookup_path) {
+        return;
+    }
+
+    push_parent_not_found(task);
+}
+
+fn push_parent_not_found(task: &mut Task) {
+    let already_exists = task.warnings.iter().any(|warning| {
+        warning.code == TaskWarningCode::ParentNotFound
+            && warning.field.as_deref() == Some("parent")
+    });
+    if already_exists {
+        return;
+    }
+
+    task.warnings.push(TaskWarning {
+        code: TaskWarningCode::ParentNotFound,
+        field: Some("parent".to_string()),
+        message: "parent task was not found".to_string(),
+    });
+}
+
 fn find_task_mut<'a>(
     cache: &'a mut HashMap<PathBuf, Task>,
     normalized: &str,
@@ -240,3 +406,7 @@ fn find_task_mut<'a>(
 #[cfg(test)]
 #[path = "task_index_tests.rs"]
 mod task_index_tests;
+
+#[cfg(test)]
+#[path = "task_index_parent_chain_tests.rs"]
+mod task_index_parent_chain_tests;
